Moves "common operators" out of pgx-utils
The only place they're seemingly used is in pgx-macros, so extracting
them out of pgx-utils and putting them into macros might be ideal.
